+++ test/test_pdf_combiner.dart
@@ -19,7 +19,7 @@ void main() {
       });
       DocumentUtils.setTemporalFolderPath("./example/assets/temp");
       var result = await PdfCombiner.generatePDFFromDocuments(
-        inputPaths: ["./example/assets/image_1.jpeg"],
+        inputs: [MergeInputPath("./example/assets/image_1.jpeg")],
         outputPath: "./example/assets/temp/document_0.pdf",
       );
       expect(result, "./example/assets/temp/document_0.pdf");